defer webpack config generation until dist-metal has been run since it depends on existence of generated html files in .tmp/
